+++ llama_index/indices/common/struct_store/base.py
@@ -9,21 +9,15 @@ from llama_index.data_structs.table import StructDatapoint
 from llama_index.indices.service_context import ServiceContext
 from llama_index.langchain_helpers.sql_wrapper import SQLDatabase
 from llama_index.llm_predictor.base import BaseLLMPredictor
-from llama_index.prompts.default_prompt_selectors import (
-    DEFAULT_REFINE_TABLE_CONTEXT_PROMPT_SEL,
-)
-from llama_index.prompts.default_prompts import (
-    DEFAULT_TABLE_CONTEXT_PROMPT,
-    DEFAULT_TABLE_CONTEXT_QUERY,
-)
+from llama_index.prompts.default_prompt_selectors import \
+    DEFAULT_REFINE_TABLE_CONTEXT_PROMPT_SEL
+from llama_index.prompts.default_prompts import (DEFAULT_TABLE_CONTEXT_PROMPT,
+                                                 DEFAULT_TABLE_CONTEXT_QUERY)
 from llama_index.prompts.prompt_type import PromptType
-from llama_index.prompts.prompts import (
-    QuestionAnswerPrompt,
-    RefinePrompt,
-    RefineTableContextPrompt,
-    SchemaExtractPrompt,
-    TableContextPrompt,
-)
+from llama_index.prompts.prompts import (QuestionAnswerPrompt, RefinePrompt,
+                                         RefineTableContextPrompt,
+                                         SchemaExtractPrompt,
+                                         TableContextPrompt)
 from llama_index.response_synthesizers import get_response_synthesizer
 from llama_index.schema import BaseNode, MetadataMode
 from llama_index.text_splitter import TextSplitter
