+++ llama_index/node_parser/node_utils.py
@@ -4,16 +4,10 @@
 import logging
 from typing import List
 
-from llama_index.schema import (
-    BaseNode,
-    Document,
-    ImageDocument,
-    ImageNode,
-    MetadataMode,
-    NodeRelationship,
-    TextNode,
-)
-from llama_index.text_splitter import TextSplit, TextSplitter, TokenTextSplitter
+from llama_index.schema import (BaseNode, Document, ImageDocument, ImageNode,
+                                MetadataMode, NodeRelationship, TextNode)
+from llama_index.text_splitter import (TextSplit, TextSplitter,
+                                       TokenTextSplitter)
 from llama_index.utils import truncate_text
 
 logger = logging.getLogger(__name__)
