+++ llama_index/node_parser/simple.py
@@ -4,7 +4,8 @@ from typing import List, Optional, Sequence
 from llama_index.callbacks.base import CallbackManager
 from llama_index.callbacks.schema import CBEventType, EventPayload
 from llama_index.constants import DEFAULT_CHUNK_OVERLAP, DEFAULT_CHUNK_SIZE
-from llama_index.node_parser.extractors.metadata_extractors import MetadataExtractor
+from llama_index.node_parser.extractors.metadata_extractors import \
+    MetadataExtractor
 from llama_index.node_parser.interface import NodeParser
 from llama_index.node_parser.node_utils import get_nodes_from_document
 from llama_index.schema import BaseNode, Document
