+++ tests/indices/test_node_utils.py
@@ -4,7 +4,8 @@ from typing import List
 
 import pytest
 
-from llama_index.node_parser.node_utils import TextSplit, get_nodes_from_document
+from llama_index.node_parser.node_utils import (TextSplit,
+                                                get_nodes_from_document)
 from llama_index.schema import Document, MetadataMode
 from llama_index.text_splitter import TokenTextSplitter
